+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.18
 - split CI into smaller jobs
+- fixed unique constraint creation with the `deferrable` parameter
 
 ## 0.17
 - added django 5.1 support